Resolve errors and refactoring
